fixed at[] index

+++ egs/egs_ref.py
@@ -169,10 +169,10 @@ class CleanTx():
         """Rounds the gas price to gwei"""
         gp = self.gas_price/1e8
         if gp >= 1 and gp < 10:
-            gp = np.ceil(gp)
+            gp = int(np.ceil(gp))
         elif gp >= 10:
             gp = gp/10
-            gp = np.ceil(gp)
+            gp = int(np.ceil(gp))
             gp = gp*10
         else:
             gp = 0
